Don't provide hulls for entities that don't collide
