+++ nova/BUILD.bazel
@@ -50,32 +50,19 @@ ts_library(
     ]
 )
 
-esbuild(
-    name = "server_bundle",
-    args = [
-        "--keep-names", # For convenience. Can be removed.
-        "--resolve-extensions=.mjs,.js",
-    ],
-    entry_point = "server.ts",
-    platform = "node",
-    minify = False,
-    sources_content = True,
-    define = [
-        "process.env.NODE_ENV=\"node\""
-    ],
-    deps = [
-        ":server",
-    ],
-)
-
 data = [
-    ":server_bundle",
+    #":server_bundle",
+    ":server",
     "//nova/settings",
     ":nova_data",
     "//nova/src:static_files",
     "//nova/src:browser_bundle",
     "//nova/src/server/parsing:nova_parse_worker_bundle",
-    #"//nova/src/engine",
+    #"//nova/src/engine_lib",
+    "//nova/src/communication:communication_lib",
+    "//nova/src/communication:communication_server_lib",
+    #"@npm//comlink",
+    #"@npm//:node_modules",
     # "//nova/src/proto:protobufjs_bundle",
     # "@npm//protobufjs",
     #"@npm_bazel_labs//protobufjs:bootstrap_scripts",
@@ -83,8 +70,10 @@ data = [
 
 nodejs_binary(
     name = "nova",
-    entry_point = "server_bundle.ts",
+    #entry_point = "server_bundle.ts",
+    entry_point = "server.ts",
     data = data,
+    link_workspace_root = True,
 )
 
 container_image(
@@ -96,9 +85,10 @@ container_image(
 
 nodejs_image(
     name = "nova_image",
-    entry_point = "server_bundle.ts",
+    entry_point = "server.ts",
     # npm deps will be put into their own layer
     data = data,
+    link_workspace_root = True,
     base = ":port_image",
 )
