+++ nova/server.ts
@@ -19,7 +19,8 @@ import { isLeft } from "fp-ts/lib/Either";
 import { makeSystem } from "./src/nova_plugin/make_system";
 import { Worker } from "worker_threads";
 import * as Comlink from 'comlink';
-import nodeEndpoint from "comlink/dist/esm/node-adapter";
+//import nodeEndpoint from "comlink/dist/esm/node-adapter";
+import nodeEndpoint from "comlink/dist/umd/node-adapter";
 import { NovaParseWorkerApi } from "./src/server/parsing/nova_parse_worker";
 import { makeShip } from "./src/nova_plugin/make_ship";
 import { v4 } from "uuid";
