+++ nova/src/BUILD.bazel
@@ -71,7 +71,7 @@ esbuild(
     ],
     entry_point = "browser.ts",
     target = "es2015",
-    minify = True,
+    minify = False,
     sources_content = True,
     define = [
         "process.env.NODE_ENV=\"browser\""
