+++ nova/src/nova_plugin/weapons_state.ts
@@ -1,0 +1,15 @@
+import * as t from 'io-ts';
+import { Component } from 'nova_ecs/component';
+import { map } from 'nova_ecs/datatypes/map';
+
+const WeaponState = t.intersection([t.type({
+    count: t.number,
+    firing: t.boolean,
+}), t.partial({
+    target: t.string,
+})]);
+export type WeaponState = t.TypeOf<typeof WeaponState>;
+
+export const WeaponsState = map(t.string /* weapon id */, WeaponState);
+export type WeaponsState = t.TypeOf<typeof WeaponsState>;
+export const WeaponsStateComponent = new Component<WeaponsState>('WeaponsStateComponent');
